+++ TeacherHelper/src/testTypeManagementForm.cpp
@@ -47,7 +47,7 @@ void TestTypeManagementForm::showEvent(QShowEvent *event)
 void TestTypeManagementForm::refreshItemsTable()
 {
 	ui.tableWidgetItems->model()->removeRows(0, ui.tableWidgetItems->rowCount());
-	size_t row {0};
+	int row {0};
     for (const auto &testType : controller.getTestTypes()) {
 		ui.tableWidgetItems->insertRow(row);
 		ui.tableWidgetItems->setItem(row, 0, new QTableWidgetItem(to_string(testType.getId()).c_str()));
